# index: Optimize generatePackList