@greatroar 2023-01-13
I also tried changing debug.Log to

```
func Log(f string, args ...interface{}) {
	if !opts.isEnabled {
		return
	}
	doLog(f, args...)
}

func doLog(f string, args ...interface{}) {
	/* ... */
}
```

While debug.Log can then be inlined, the allocation still happens.