@MichaelEischer 2023-01-14
> While debug.Log can then be inlined, the allocation still happens.

I guess the escape analysis isn't able to fully analyze the vararg. If I remember correctly, then the allocation should even happen, if `Log` is empty.