@greatroar 2023-01-15
Ugh. If I remove restic.ID.MarshalJSON, I can get even fewer allocations, but then encoding indexes gets slower:

```
name                 old time/op    new time/op    delta
EncodeIndex/100-8      1.48ms ± 3%    1.64ms ± 3%  +10.91%  (p=0.000 n=10+10)
EncodeIndex/1000-8     13.1ms ± 2%    15.9ms ± 8%  +21.08%  (p=0.000 n=10+10)
EncodeIndex/10000-8     116ms ± 2%     147ms ± 3%  +26.83%  (p=0.000 n=10+10)

name                 old alloc/op   new alloc/op   delta
EncodeIndex/100-8       275kB ± 1%     193kB ± 2%  -29.73%  (p=0.000 n=10+10)
EncodeIndex/1000-8     2.88MB ± 5%    2.46MB ±13%  -14.54%  (p=0.000 n=9+10)
EncodeIndex/10000-8    31.9MB ±10%    29.2MB ±25%     ~     (p=0.190 n=10+10)

name                 old allocs/op  new allocs/op  delta
EncodeIndex/100-8       2.39k ± 0%     1.38k ± 0%  -42.13%  (p=0.000 n=10+10)
EncodeIndex/1000-8      22.9k ± 0%     13.3k ± 0%  -42.18%  (p=0.000 n=9+10)
EncodeIndex/10000-8      229k ± 0%      132k ± 0%  -42.28%  (p=0.000 n=10+10)
```

This warrants some more research, but I'll leave that out of this PR.